Consolidate header and hide when release notes are displayed
